Add a simple monitoring loop component that monitors the cluster
The monitor checks both edge transitions and periodically samples
conditions on the cluster to give the test framework context about
the actions happening on the cluster when a test fails, as well as
allowing a dump of the entire timeline.

Expose monitoring as a test subcommand for debugging, and wire it
up to be invoked when a test fails.

Skip printing successful test output by default, include time in message

Print the time a test completed in the lines. Support a new argument
`--include-success` if the user wants to receive output from successful
tests.