test/extended: Use internal service if no router LB
If the ingress controller/router does not have a LoadBalancer service, use
the internal service for tests.

Delete broken skip logic.  The logic in question was checking an error
value from wait.PollImmediate using the IsNotFound function from the
Kubernetes API machinery.  In the polling in question, if a resource is not
found, the polling eventually times out, and wait.PollImmediate returns
wait.ErrWaitTimeout, which is not an APIStatus type, which means that
IsNotFound would never return true for it.

This commit fixes bug 1700068.

https://bugzilla.redhat.com/show_bug.cgi?id=1700068

* test/extended/router/router.go
(waitForRouterMetricsIP): Rename from this...
(waitForRouterInternalIP): ...to this.
(waitForRouterServiceIP): Rename from this...
(waitForRouterExternalIP): to this.
(routerShouldHaveExternalService): New function that checks the default
ingresscontroller's status to determine whether it is supposed to be
exposed by an LB service.
(waitForRouterServiceIP): New function that uses
routerShouldHaveExternalService and returns the result of either
waitForRouterInternalIP or waitForRouterExternalIP accordingly.

* test/extended/util/cli.go (AdminOperatorClient): New method (used by
routerShouldHaveExternalService).

* test/extended/router/router.go:
* test/extended/router/headers.go:
* test/extended/router/metrics.go:
* test/extended/router/reencrypt.go: Delete broken skip logic.
Use waitForRouterServiceIP when some address is needed.
Use waitForRouterInternalIP when an internal address is needed.