Allow skipping tests based on detected network plugin
Detect the network plugin (and, for openshift-sdn, the mode) and allow
skip lists to be configured based on that.

Remove the openshift/conformance/multitenant test suite since it will
no longer be needed.

Add a list of tests to be skipped for ovn-kubernetes in 4.2 due to
not-yet-implemented upstream functionality.

Move the skipped unimplemented-NetworkPolicy-feature tests out of the
"global" skip list and into the openshift-sdn-specific list, since
they should still be run for ovn-kubernetes, kuryr, and third-party
plugins.