test: Skip unpacking file contents except when the test is actually running
In general test startup should do a minimal amount of static work before
the test suite runs. Set up guard rails within exutil that prevent tests
from running commands. Standardize the setup of the test process now that
key methods on the kube Framework are public. Standardize how exutil is
included and called in all packages.

The command test is switched to asking for a list of filenames from the
fixtures. The multicast test no longer instantiates a framework outside
of a block. Ensure that fixtures are cleaned up after the test exits
instead of being left around.