oVirt: skip network service tests
This PR skips the network service tests for ovirt.
we didn't know the reason until recently and now there is a fix which is targeted for RHEL 8.3 [1] and there are openshift bugs for tracking the fix in openshift [2][3][4].
Since this is now a very known issue and is currently failing our CI and making it hard to distinguish between a red failed CI and a false one we decided to skip those test for ovirt until the fix will land.

Tracker-Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1838751

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1794714
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1822714
[3] https://bugzilla.redhat.com/show_bug.cgi?id=1822755
[4] https://bugzilla.redhat.com/show_bug.cgi?id=1822766

Signed-off-by: Gal-Zaidman <gzaidman@redhat.com>