Monitor availability of registry through router
When the tests are run outside of the cluster (usually they are), the
tests cannot access the registry service. They also cannot rely on the
registry route, because the registry route is disabled by default.

To check availability of the registry, tests can create additional route
for the registry service and monitor it during upgrades.