test/extended/prometheus: Validate alerting rules
The [OpenShift Alerting Consistency][1] enhancement defines a style
guide for the alerts shipped as part of OpenShift.  This adds a test
validating some of the guidelines considered required.

This was originally added in #26476, but was reverted in #26499 due to
failures with OVN and vSphere clusters.  This adds the tests back, but
adds exceptions for the non-compliant alerts as well as marking the
failing tests as flakes for now.  We'll gather data and make the tests
required once we're reasonably sure things are passing with all the
existing alerts.

[1]: https://github.com/openshift/enhancements/blob/master/enhancements/monitoring/alerting-consistency.md