test: skip tests that won't work behind a proxy automatically
The vast majority of the tests we run work and run fine through
an HTTP proxy to reach the target cluster under test, however a
handful of tests don't work. This adds a configuration to
openshift-tests to see if given the current proxy environment, we
would use a proxy to reach the API server. If so, then we skip
tests annotated with `[Skip:Proxy]`.