Tolerate API downtime in upgrade disruption tests on single-node topologies
Currently the tests tolerate 8% API disruption on <4.8 clusters and 0% API disruption on
>=4.8 azure,aws,gce clusters

This test makes it so no matter what version / cloud provider is used, if the cluster control
plane topology is single node, then 15% API disruption is allowed during an upgrade. This is
because there is no way to avoid disruption on single-node topologies, but we would still
like to make sure that we don't regress to too much downtime.