Extutil: Add helpers to get a Hypershift management CLI
This adds some helpers to get a management cluster CLI for Hypershift clusters.
Contrary to a normal CLI it does not perform any teardown, which is
mostly the result of Ginkgo not allowing to register post handlers
inside a test (It() block) and also not allowing to perform assertions
outside a test. As finding out if we are testing a hypershift cluster
can error, it has to happen inside a test which means that retrieving
the management cluter CLI also has to happen there, because we have to
perform that check to see if we need it.