test: avoid double slash in path
every user of configObjectsDir already appends a slash after
configObjectsDir.  Avoid generating a path that has a double slash
as `var/oauth/configobjects//htpasswd`.

The reason for such fix is that the crun version currently used in the
CI has a bug when dealing with paths with multiple slashes.  The issue
has already been addressed upstream.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>