Bump timeout for CVO to ack the upgrade generation.
Team thinks this is related to verification of the payload, and is
affecting only a few metal jobs for the most part, but doing so
very reliably.

Proper investigation and fix being pursued in
https://bugzilla.redhat.com/show_bug.cgi?id=2071998

If we're over the default 2 mins, but under 4 on those platforms, we'll
flake the test rather than just pass.