test/extended/util/openshift/clusterversionoperator: Survive API hiccups
The test case:

  [sig-cluster-lifecycle] TestAdminAck should succeed [Suite:openshift/conformance/parallel]

is vulnerable to brief API-server hiccups like [0,1]:

  Dec  6 00:00:10.440: FAIL: Error accessing configmap openshift-config-managed/admin-gates: Get "https://api.ci-op-g2m38jp7-eafe9.origin-ci-int-aws.dev.rhcloud.com:6443/api/v1/namespaces/openshift-config-managed/configmaps/admin-gates": dial tcp: lookup api.ci-op-g2m38jp7-eafe9.origin-ci-int-aws.dev.rhcloud.com on 172.30.0.10:53: no such host

and [2]:

  Dec  9 19:53:20.747: FAIL: Error accessing configmap openshift-config-managed/admin-gates: Get "https://api.ci-op-w5q90zpi-9278e.origin-ci-int-aws.dev.rhcloud.com:6443/api/v1/namespaces/openshift-config-managed/configmaps/admin-gates": dial tcp 100.21.251.165:6443: i/o timeout

This commit logs the error and continues to the next poll, in case
that works.  We keep track of these "I couldn't get far enough in to
confirm whether this works or not" issues, and complain about the most
recent if we reach the end of the run and have never successfully
passed a test-attempt.

There's still some false-failure exposure risk if we fail in
setAdminGate or one of those steps, but the exposure is "once per ack
being excercised" which is a lot smaller than the initial gets' "twice
per poll attempt".

There's also a risk of false-success, if we pass on an early version
(e.g. "Admin ack is not in this baseline or contains no gates"), but
then fail to reach the Kube API for the rest of the run after updating
to some later version.  But I expect situations where our poller
repeatedly fails to connect to the Kube API server without ever
recovering while we're on a particular version will be rare, and that
some other Kube-API reachability test-case will fail the overall
suite.

[0]: https://bugzilla.redhat.com/show_bug.cgi?id=2030972
[1]: https://bugzilla.redhat.com/show_bug.cgi?id=2026806#c8
[2]: https://bugzilla.redhat.com/show_bug.cgi?id=2027929#c1