Add disruption test to an external service, not cluster under test.
TRT has identified situations where disruption testing fails for a
batch of jobs talking to clusters in many platforms, all at the same
time, when the ci jobs were all running on same build cluster.

To help identify when we're seeing real disruption of the cluster under
test, vs the CI cluster itself losing network, we've stood up a simple
service in DPCR cluster, and added a new disruption backend that will
contiually poll that endpoint. If we see disruption here, we know any
disruption of the cluster under test is likely not real.

In future this may expand to automatically pass disruption tests that
fail if we see corresponding network problems for the CI cluster.