Remove etcd vertical scaling test.
This test is extremely disruptive to the cluster and while lots of work
was done to try to keep invariant tests happy, it has continued to
surface problems that are failing jobs and blocking payloads and code
shipping to the org.

After discussion with etcd team we're agreed this is a bad fit for the
serial suite and will continue to cause problems for any future
invariant checks we add. It should be moved to separate jobs and run on
etcd related pull requests and possibly as a periodic.

Removing to help improve aws-serial pass rates.

TRT-358