Ensures operators use external cloud provider
OpenStack platform is moving from in-tree cloud provider
to external one. The switch to external cloud provider has
to be done at the same time for the cluster cloud controller
manager operator and kube controller manager operator to avoid
the two providers being used in the cluster. Besides this the
machine config operator has to configure kubelet with external provider.
This commit includes a new test scenario to check that external provider
is properly configured for the needed operators.

Co-authored-by: Joel Speed <Joel.speed@hotmail.co.uk>