Add a test case to find job name/cluster version mismatches
We have at least one case of a job saying it's 4.11 but actually
deploying 4.12. I don't think we can get a perfect test to cover this,
but for periodics we can at least check that if a job name mentions a
version, that exists in the cluster version history.

There are jobs that mention versions that aren't openshift which will be
problematic, but this is just flaking for now until we figure out how to
deal with those.